+++ test/Main.hs
@@ -8,6 +8,7 @@ import qualified Test.Bitcask.Crash
 import qualified Test.Bitcask.Keydir
 import qualified Test.Bitcask.Model
 import qualified Test.Bitcask.Record
+import qualified Test.Bitcask.WriteSafety
 
 main :: IO ()
 main =
@@ -25,5 +26,6 @@ main =
           [ Test.Bitcask.Model.tests
           , Test.Bitcask.Crash.tests
           , Test.Bitcask.Concurrency.tests
+          , Test.Bitcask.WriteSafety.tests
           ]
       ]